add checking if file already exist

+++ environment/development/setup.sh
@@ -1,4 +1,7 @@
 #!/bin/bash
 DIR="$( cd "$( dirname "$0" )" && pwd )"
-ln -fs $DIR/githooks/pre-commit $DIR/../../.git/hooks/pre-commit
+if ! [ -e "$DIR/../../.git/hooks/pre-commit" ]
+then
+	ln -fs $DIR/githooks/pre-commit $DIR/../../.git/hooks/pre-commit
+fi
